fix: adjust regional ID types and security permissions for manual sync

+++ backend/src/main/java/com/bruno/artistalbum/model/Regional.java
@@ -14,7 +14,7 @@ public class Regional {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long internalId;
+    private Integer internalId;
 
     private Integer id; // ID que vem do endpoint
